+++ src/App.js
@@ -1,5 +1,5 @@
 import { Component } from 'react';
-import ListNews from './components/ListNews'
+import ListNews from './components/main-news/ListNews'
 import { dataBase } from './Data';
 
 class App extends Component {
