--- src/components/CardNew.jsx
@@ -1,39 +1,0 @@
-import { Component } from "react";
-
-class CardNew extends Component {
-    render() {
-        return (
-            <div className="body-card">
-
-                <div className="card-text">
-
-                    <p className="description-card-text">
-                        El 'Infojobs del Gobierno' ya es oficial: esto es todo lo que sabemos de este nuevo portal de empleo estatal
-                    </p>
-
-                    <p className="categoria-card-text">
-                        Empresas y Economia
-                    </p>
-
-                    <p className="cuerpo-card-text">
-                        El Consejo de Ministros ha a aprobado este viernes el anteproyecto de ley que regulara este servicio. <a href="#" className="read-more">Leer Mas</a>
-                    </p>
-
-                    <p className="footer-card-text">
-                        <a href="#" className="comments-card-text">💬</a> - Pablo Rodriguez - Hace 39 Min"
-                    </p>
-
-                </div>
-
-                <div className="card-img">
-
-                    <img src="https://i.blogs.es/a2a76c/1366_2000-37-/500_333.jpg" alt="#" />
-
-                </div>
-
-            </div>
-        )
-    }
-}
-
-export default CardNew
