+++ src/components/main-news/ListNews.jsx
@@ -1,11 +1,14 @@
 import { Component } from "react";
 import CardNew from './CardNew';
+import './styles.css';
 
 class ListNews extends Component {
     render() {
         return (
-            <div className="main-list">
-                <CardNew />
+            <div className="main-content">
+                <div className="main-list">
+                    <CardNew />
+                </div>
             </div>
         )
     }
