+++ src/components/main-news/styles.css
@@ -1,0 +1,66 @@
+/* Estilo provisional */
+.main-content{
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    height:100%;
+    background-color: rgb(199, 195, 195);
+}
+
+/* Lista de cardsnews*/
+.main-list{
+    width: 75%;
+    height: 80%;
+    margin: auto;
+    box-shadow: 0 0 5px white;
+}
+
+.body-card{
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    width: 100%;
+    height: 40%;
+    box-shadow: 0 0 5px white;
+}
+
+.card-img{
+    display: flex;
+    margin: auto;
+    width: 40%;
+    height: 100%;
+}
+
+.card-img img{
+    object-fit: cover;
+    height: 100%;
+    width: 100%;
+}
+
+.card-text{
+    display: flex;
+    flex-direction: column;
+    padding: 10px;
+    gap: 10%;
+    width: 60%;
+    font-family: Verdana, Geneva, Tahoma, sans-serif;
+    border: 1px solid red;
+}
+
+.description-card-text{
+    font-size: 20px;
+    font-weight: bolder;
+}
+
+.categoria-card-text{
+    font-size: 15px;
+    text-transform: uppercase;
+    color: green;
+    font-weight: lighter;
+}
+
+.cuerpo-card-text{
+
+}
