+++ src/general.css
@@ -1,0 +1,9 @@
+*{
+    margin: 0;
+    padding: 0;
+}
+
+html, body, #root{
+    width: 100%;
+    height: 100%;
+}
