+++ src/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
+import './general.css';
 
 ReactDOM.render(
   <React.StrictMode>
